fix(ui): keep component files free of non-component exports
react-doctor fails the build on `only-export-components`, and four exports
tripped it: `fromLucide` sat beside the 57 lucide re-exports, and menu.tsx
shipped two class helpers and `explainAction` alongside its components. Both
split in two -- `adapters/lucide-symbols.ts` and `menu/menu-parts.ts` -- which
keeps the icon library boundary and the shared menu surface exactly where they
were.

The nine remaining findings are all in stories, where `args: { children }` is
how Storybook declares args and `render={<h3 />}` is the element Base UI fills
in later. Both rules are switched off for `*.stories.tsx` and stay on
everywhere else.
